+++ test/logic/helper/api_cancel_token_test.dart
@@ -36,12 +36,12 @@ void main() {
       expect(token.isClosed, false);
     });
 
-    test("Can cancel", () {
+    test("Cancel and removed cancellable", () {
       final token = ApiCancelToken();
       token.attachCancellable(CancelableOperation.fromValue(1));
       token.close();
       expect(token.isClosed, true);
-      expect(token.haveCancellables, true);
+      expect(token.haveCancellables, false);
     });
 
     test("Can close empty", () {
@@ -56,12 +56,45 @@ void main() {
       final operation = CancelableOperation.fromFuture(
         Future.delayed(const Duration(seconds: 2), () => 1),
       );
+
       token.attachCancellable(operation);
       Future(() => token.close());
+
       await expectLater(
         operation.then((p0) => 0, onCancel: () => throw Exception()).value,
         throwsException,
       );
+    });
+
+    test("Closing several futures", () async {
+      final token = ApiCancelToken();
+      final operation1 = CancelableOperation.fromFuture(
+        Future.delayed(const Duration(seconds: 2), () => 1),
+      );
+      final operation2 = CancelableOperation.fromFuture(
+        Future.delayed(const Duration(seconds: 2), () => 1),
+      );
+
+      token.attachCancellable(operation1);
+      token.attachCancellable(operation2);
+      token.close();
+
+      expect(operation1.isCanceled, true);
+      expect(operation2.isCanceled, true);
+    });
+
+    test("Add cancelable after close", () async {
+      final token = ApiCancelToken();
+      final operation = CancelableOperation.fromFuture(
+        Future(() => 1),
+      );
+
+      token.close();
+      token.attachCancellable(operation);
+
+      expect(token.isClosed, true);
+      expect(token.haveCancellables, false);
+      expect(operation.isCanceled, false);
     });
   });
 }